requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.0.8
django-cors-headers==3.5.0
djangorestframework==3.11.0
gunicorn==20.0.4
mysqlclient==2.0.1
Pillow==7.2.0
psycopg2==2.8.6
python-decouple==3.3
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.2.0
